fixed: Change Directory Architecture
