+++ app/components/CustomReactMarkdown.tsx
@@ -3,7 +3,7 @@ import ReactMarkdown from 'react-markdown'
 import remarkGfm from 'remark-gfm'
 import Image from 'next/image'
 import { ClassNames } from '@emotion/react'
-import SimpleButton from './CustomButton'
+import SimpleButton from './ui/button/SimpleButton'
 
 interface CustomReactMarkdownProps {
     content: string,
