+++ app/components/ui/decoration/OneBlueLine.tsx
@@ -1,0 +1,9 @@
+const OneBlueLine = () => {
+    return (
+        <div className="relative origin-top -rotate-12 relative m-0 py-10 w-full bg-primary-400 z-0">
+                <div className=""></div>
+        </div>
+    )
+}
+
+export default OneBlueLine
