+++ app/components/ui/decoration/SingleYellowLines.tsx
@@ -1,0 +1,7 @@
+const SingleYellowLines = () => {
+    return (
+        <div className="absolute bg-secondary-200 h-screen w-screen opacity-50 -z-10" style={{clipPath: "polygon(0% 100%, 0% 80%, 100% 0%, 100% 20%)"}}></div>
+    )
+}
+
+export default SingleYellowLines
