+++ app/components/ui/infoCard/InfoCard.tsx
@@ -28,7 +28,7 @@ const InfoCard: React.FC<InfoCardProps> = ({ category, id, title, description, t
     if (title.length > 30) title = title.slice(0, 30) + '...';
     if (description.length > 55) description = description.slice(0, 55) + '...';
 
-    if (thumbnaile === "") thumbnaile = "/img/noimage.png";
+    if (thumbnaile === "") thumbnaile = "/img/other/noimage.png";
 
     const _time = time.slice(0, 10);
 
@@ -52,54 +52,6 @@ const InfoCard: React.FC<InfoCardProps> = ({ category, id, title, description, t
 };
 
 export default InfoCard;
-
-
-export const MiniInfoCard: React.FC<{ id: number, title: string, thumbnaile: string, time: string }> = ({ id, title, thumbnaile, time }) => {
-
-
-    if (thumbnaile === "") thumbnaile = "/img/noimage.png";
-
-    let category = "event";
-
-    // titleの識別子の削除
-    if (title.includes('!!!')) title = title.replace('!!!', '');
-    if (title.includes('???')) {
-        title = title.replace('???', '');
-        category = "news";
-    }
-
-    // titleに@@が含まれている場合@を削除する（開催中のイベント）
-    if (title.includes('@@')) {
-        title = title.replace('@@', '');
-    }
-
-    if (title.length > 9) title = title.slice(0, 9) + '...';
-
-    return (
-        <>
-            <Link href={`/${category}/article/${id}`} className='' legacyBehavior>
-                <div className='w-40 md:w-48 hover:animate-pulse'>
-                    <Image src={thumbnaile} alt="Image" sizes="(max-width: 1000px) 100vw" className="w-full" width={300} height={300} />
-                    <div className='w-full'>
-                        <a className="text-base md:text-lg font-bold text-black text-right mb-2" >{title}</a>
-                    </div>
-                </div>
-            </Link>
-        </>
-    )
-}
-
-export const TopInfoCard: React.FC<{ id: number, thumbnaile: string }> = ({ id, thumbnaile }) => {
-
-    if (thumbnaile === "") thumbnaile = "/img/noimage.png";
-    return (
-        <>
-            <Link href={`/event/article/${id}`} legacyBehavior>
-                <Image src={thumbnaile} alt="Image" sizes="(max-width: 900px) 90vw" className="w-full hover:animate-pulse" width={600} height={600} />
-            </Link>
-        </>
-    )
-}
 
 
 // 開催中か終了かを示すバッジ
