+++ app/faq/page.tsx
@@ -5,7 +5,7 @@ import { url } from '@/app/layout';
 
 const title = "よくあるご質問";
 const description = "工学院ハッカソンに関するよくあるご質問をまとめました．";
-const image = "/img/ogp_faq.png";
+const image = "/img/ogp/ogp_faq.png";
 export const metadata: Metadata = {
     title: title,
     description: description,
