+++ app/whats-hackathon/page.tsx
@@ -6,7 +6,7 @@ import { url } from '@/app/layout';
 
 const title = "ハッカソンとは？";
 const description = "ハッカソンという言葉を聞いたことのない人向けに解説します．";
-const image = "/img/ogp_whats_hackathon.png";
+const image = "/img/ogp/ogp_whats_hackathon.png";
 export const metadata: Metadata = {
     title: title,
     description: description,
